backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.55
langchain==0.3.7
langchain-groq==0.2.1
langchain-core==0.3.15
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
python-multipart==0.0.12
